+++ example/lib/main.dart
@@ -8,20 +8,54 @@ void main() {
 // Define tournaments data
 final List<Tournament> _tournaments = [
   Tournament(matches: [
-    TournamentMatch(id: "1", teamA: "Real Madrid", teamB: "Barcelona", scoreTeamA: "3", scoreTeamB: "1"),
-    TournamentMatch(id: "2", teamA: "Chelsea", teamB: "Liverpool", scoreTeamA: "0", scoreTeamB: "1"),
-    TournamentMatch(id: "3", teamA: "Juventus", teamB: "Paris Saint-Germain", scoreTeamA: "0", scoreTeamB: "2"),
-    TournamentMatch(id: "4", teamA: "Manchester City", teamB: "Inter Milan", scoreTeamA: "4", scoreTeamB: "2"),
+    TournamentMatch(
+        id: "1",
+        teamA: "Real Madrid",
+        teamB: "Barcelona",
+        scoreTeamA: "3",
+        scoreTeamB: "1"),
+    TournamentMatch(
+        id: "2",
+        teamA: "Chelsea",
+        teamB: "Liverpool",
+        scoreTeamA: "0",
+        scoreTeamB: "1"),
+    TournamentMatch(
+        id: "3",
+        teamA: "Juventus",
+        teamB: "Paris Saint-Germain",
+        scoreTeamA: "0",
+        scoreTeamB: "2"),
+    TournamentMatch(
+        id: "4",
+        teamA: "Manchester City",
+        teamB: "Inter Milan",
+        scoreTeamA: "4",
+        scoreTeamB: "2"),
   ]),
   Tournament(matches: [
-    TournamentMatch(id: "1", teamA: "AC Milan", teamB: "Atletico Madrid", scoreTeamA: "4", scoreTeamB: "0"),
-    TournamentMatch(id: "2", teamA: "Borussia Dortmund", teamB: "Tottenham Hotspur", scoreTeamA: "2", scoreTeamB: "1"),
+    TournamentMatch(
+        id: "1",
+        teamA: "AC Milan",
+        teamB: "Atletico Madrid",
+        scoreTeamA: "4",
+        scoreTeamB: "0"),
+    TournamentMatch(
+        id: "2",
+        teamA: "Borussia Dortmund",
+        teamB: "Tottenham Hotspur",
+        scoreTeamA: "2",
+        scoreTeamB: "1"),
   ]),
   Tournament(matches: [
-    TournamentMatch(id: "1", teamA: "Ajax", teamB: "Sevilla", scoreTeamA: "4", scoreTeamB: "3"),
+    TournamentMatch(
+        id: "1",
+        teamA: "Ajax",
+        teamB: "Sevilla",
+        scoreTeamA: "4",
+        scoreTeamB: "3"),
   ])
 ];
-
 
 /// The main application widget.
 class MyApp extends StatelessWidget {
@@ -45,6 +79,9 @@ class MyApp extends StatelessWidget {
             itemsMarginVertical: 20.0,
             cardWidth: 220.0,
             cardHeight: 100,
+            lineWidth: 80,
+            lineThickness: 5,
+            lineBorderRadius: 12,
             lineColor: Colors.green,
           ),
         ),